+++ CHANGELOG.md
@@ -8,16 +8,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Changed
-- **Breaking** `Picker` now requires new `Render` implementation to describe how a given type is rendered on the screen.
+- **Breaking** `Picker` now requires new `Render` implementation to describe how a given type is displayed on the screen.
   - `Picker::new` signature has changed.
   - `PickerOptions::picker` signature has changed.
+- **Breaking** `PickerOptions::query` and `Picker::set_query` now accept any argument which is `Into<String>` instead of `ToString`.
 - **Breaking** `Picker::pick` uses STDERR instead of STDOUT for interactive screen. A lock is acquired to STDERR for rendering performance in case of Mutex contention.
   - If your application requires debug logging, it is probably best to log to a file instead.
 - **Breaking** `Picker::injector` now returns a `nucleo_picker::Injector` instead of a `nucleo::Injector`. The `nucleo_picker::Injector` no longer exposes the internal match object; instead, rendering is done by the new `Render` trait.
 - User CTRL-C during `Picker::pick` now returns `io::Error` with custom error message.
 
 ### Removed
-- Suggested support for multiple columns has now been removed.
+- Suggested support for multiple columns has now been removed (multiple columns were never supported internally).
 
 ### Fixed
 - Picker no longer blocks STDIN and STDOUT.
@@ -26,7 +27,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Match highlighting
-- Robust Unicode and mutiline support
+- Robust Unicode and multi-line support
   - Correctly renders multi-line items
   - Unicode width computations to correctly handle double-width and zero-width graphemes.
 - Convenient `Render` implementations in new `render` module.