+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.6.0] - 2024-12-01
 
 ### Changed
 - **Breaking** `Picker` now requires new `Render` implementation to describe how a given type is displayed on the screen.
@@ -41,7 +41,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - `Picker` and `Injector` now implement `Extend` for convenient item adding.
   - With the optional `serde` feature, an `&Injector` now implements `DeserializeSeed` to allow adding items from the picker directly from a deserializer.
 
-## [0.5.0] - 2024-11-7
+## [0.5.0] - 2024-11-07
 
 ### Added
 - Better exposure of nucleo internals: